trend_analyser/requirements.txt
groq
pyyaml
beautifulsoup4
python-dotenv
openpyxl
schedule